Move trace comparison test to JVM only
