+++ arrow-libs/core/arrow-core/src/commonTest/kotlin/arrow/core/raise/TraceSpec.kt
@@ -1,7 +1,6 @@
 package arrow.core.raise
 
 import arrow.core.right
-import arrow.typeclasses.Semigroup
 import io.kotest.assertions.throwables.shouldThrow
 import io.kotest.core.spec.style.StringSpec
 import io.kotest.matchers.shouldBe
@@ -29,23 +28,6 @@ class TraceSpec : StringSpec({
           traced({ throw error }) { unreachable() }
         }
       }.message shouldBe msg
-    }
-  }
-
-  "Can trace a typed error" {
-    either<RuntimeException, Nothing> {
-      traced({ raise(RuntimeException("")) }) { traced ->
-        // Remove first 2 lines:
-        // arrow.core.raise.RaiseCancellationException
-        //	at arrow.core.raise.DefaultRaise.raise(Fold.kt:187)
-        val trace = traced.stackTraceToString().lines().drop(2)
-
-        // Remove first line:
-        // java.lang.RuntimeException:
-        val exceptionTrace = traced.raised.stackTraceToString().lines().drop(1)
-
-        trace shouldBe exceptionTrace
-      }
     }
   }
