+++ arrow-libs/core/arrow-core/src/nativeMain/kotlin/arrow/core/raise/CancellationExceptionNoTrace.kt
@@ -2,4 +2,4 @@ package arrow.core.raise
 
 import kotlin.coroutines.cancellation.CancellationException
 
-public actual open class CancellationExceptionNoTrace : CancellationException("Raised Continuation")
+public actual open class CancellationExceptionNoTrace : CancellationException(RaiseCancellationExceptionCaptured)
